requirements.txt
fastapi
uvicorn
python-dotenv
requests
numpy
reportlab
